requirements.txt
Flask==3.1.2
keras==3.10.0
pillow==11.3.0
tensorflow==2.16.2